% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{getToken}
\alias{getToken}
\title{Get the API token}
\usage{
getToken(apikey, apisec, url_endpoint)
}
\arguments{
\item{apikey}{API Key}

\item{apisec}{API Secret}

\item{url_endpoint}{Microclim Authentication endpoint}
}
\value{
API token (valid for 2 hours)
}
\description{
API wrapper to interact with microclim.org.
}
\details{
Token is 167 characters in length, it has the user id encoded as part of it
}
\examples{
\dontrun{
mtoken = getToken('07d4d584c04941a25e291feb8881c685','9ef6bbb24a855fbb765f3890e05592f4','localhost:3000/')
}

}
\author{
Aji John
}
\keyword{API}
\keyword{Microclim}
