% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicroclimAPI.R
\docType{class}
\name{MicroclimAPI-class}
\alias{MicroclimAPI-class}
\alias{MicroclimAPI}
\title{Reference Class for Microclim.org Interaction}
\description{
API with all the REST services
}
\details{
Each service is implemented
}

\examples{
\dontrun{
mtoken = getToken('07d4d584c04941a25e291feb8881c685','9ef6bbb24a855fbb765f3890e05592f4','localhost:3000/')
ma <- MicroclimAPI$new(token = mtoken,url_mc='http://localhost:3000/')
apis= ma$apis()
jsonlite::prettify(apis)
}

}
\author{
Aji John
}
\keyword{API}
\keyword{Microclim}
