% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicroclimRequest.R
\docType{class}
\name{MicroclimRequest-class}
\alias{MicroclimRequest-class}
\alias{MicroclimRequest}
\title{Reference Class for MicroclimRequest for Microclim.org Interaction}
\description{
Reference class encapsulating Microclim Request
}
\details{
MicroclimRequest
}

\examples{

\dontrun{
mr <- MicroclimRequest$new(
                           latS = "39.40012200014591",
                           latN="39.92132255884663",
                           lonW="-106.47674560546875",
                           lonE="-105.92193603515625",
                           variable="ALBEDO",
                           shadelevel=0,
                           hod=0,
                           interval=0,
                           aggregation=0,
                           stdate="19810101",
                           eddate="19810128",
                           file="csv")
}

}
\author{
Aji John
}
\keyword{API}
\keyword{Microclim}
